session check for instructor

+++ src/main/java/com/sk/skala/quizapi/service/InstructorService.java
@@ -1,5 +1,7 @@
 package com.sk.skala.quizapi.service;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.domain.Page;
@@ -7,25 +9,39 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 
+import com.sk.skala.quizapi.config.Constant;
 import com.sk.skala.quizapi.config.Error;
+import com.sk.skala.quizapi.data.common.AccountInfo;
 import com.sk.skala.quizapi.data.common.PagedList;
 import com.sk.skala.quizapi.data.common.Response;
 import com.sk.skala.quizapi.data.table.Instructor;
+import com.sk.skala.quizapi.data.table.Subject;
 import com.sk.skala.quizapi.exception.ParameterException;
 import com.sk.skala.quizapi.exception.ResponseException;
 import com.sk.skala.quizapi.repository.InstructorRepository;
+import com.sk.skala.quizapi.repository.SubjectRepository;
+import com.sk.skala.quizapi.tools.JwtTool;
 import com.sk.skala.quizapi.tools.SecureTool;
 import com.sk.skala.quizapi.tools.StringTool;
 
+import jakarta.servlet.http.Cookie;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class InstructorService {
 	private final InstructorRepository instructorRepository;
+	private final SubjectRepository subjectRepository;
+	private final SessionHandler sessionHandler;
 
 	public Response getInstructorList(String name, int offset, int count) throws Exception {
+		if (!sessionHandler.isAdmin()) {
+			throw new ResponseException(Error.NOT_AUTHORIZED);
+		}
+
 		Pageable pageable = PageRequest.of(offset, count, Sort.by(Sort.Direction.ASC, "instructorName"));
 		Page<Instructor> paged = instructorRepository.findAllByInstructorNameLike(StringTool.like(name), pageable);
 
@@ -56,17 +72,24 @@ public class InstructorService {
 		if (!password.equals(item.getInstructorPassword())) {
 			throw new ResponseException(Error.INVALID_ID_OR_PASSWORD);
 		}
-		instructor.setInstructorPassword("");
+
+		AccountInfo account = storeAccessToken(instructor);
 
 		Response response = new Response();
-		response.setBody(instructor);
+		response.setBody(account);
 
 		return response;
 	}
 
 	public Response upsertInstructor(Instructor item) throws Exception {
+
 		if (StringTool.isAnyEmpty(item.getInstructorEmail(), item.getInstructorName(), item.getInstructorPassword())) {
 			throw new ParameterException("instructorEmail", "instructorName", "instructorPassword");
+		}
+
+		AccountInfo account = sessionHandler.getAccountInfo();
+		if (account == null || !account.getAccountId().equalsIgnoreCase(item.getInstructorEmail())) {
+			throw new ResponseException(Error.NOT_AUTHORIZED);
 		}
 
 		Optional<Instructor> option = instructorRepository.findByInstructorEmail(item.getInstructorEmail());
@@ -82,6 +105,10 @@ public class InstructorService {
 	}
 
 	public Response deleteInstructor(Instructor item) {
+		if (!sessionHandler.isAdmin()) {
+			throw new ResponseException(Error.NOT_AUTHORIZED);
+		}
+
 		if (instructorRepository.existsById(item.getId())) {
 			instructorRepository.deleteById(item.getId());
 		} else {
@@ -89,4 +116,28 @@ public class InstructorService {
 		}
 		return new Response();
 	}
+
+	public AccountInfo storeAccessToken(Instructor instructor) {
+		AccountInfo account = new AccountInfo(instructor.getInstructorEmail(), instructor.getInstructorName(),
+				AccountInfo.ROLE_WRITER);
+
+		List<Subject> subjects = subjectRepository.findAllByInstructorId(instructor.getId());
+		List<Long> subjectIds = new ArrayList<>();
+		for (Subject subject : subjects) {
+			subjectIds.add(subject.getId());
+		}
+		account.setSubjectIds(subjectIds);
+
+		String token = JwtTool.generateToken(account.getAccountId(), account, Constant.JWT_SECRET_BFF);
+		Cookie cookie = new Cookie(Constant.JWT_ACCESS_COOKIE, token);
+		cookie.setMaxAge(Constant.JWT_ACCESS_TTL);
+		cookie.setPath("/");
+		cookie.setSecure(false);
+
+		ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
+		attr.getResponse().addCookie(cookie);
+
+		return account;
+	}
+
 }
